backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
SQLAlchemy==2.0.22
pymysql==1.0.3
alembic==1.12.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic-settings==2.0.3
cryptography==42.0.5
pydantic[email]==2.0.3